Dropped SPI for tasks, moved interfaces into webserver package.
