+++ nima/webserver/webserver/src/main/java/io/helidon/nima/webserver/HelidonTaskExecutor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.nima.webserver.task.spi;
+package io.helidon.nima.webserver;
 
 import java.io.Closeable;
 import java.util.concurrent.Future;
@@ -25,7 +25,7 @@ import java.util.concurrent.TimeUnit;
  * {@link InterruptableTask}s and can be efficiently terminated. A thread that is
  * waiting to read on an open connection cannot be efficiently stopped. This
  * executor will query the thread and interrupt it if possible. It is important
- * to efficiently shut down the Nima webserver efficiently certain environments.
+ * to efficiently shut down the Nima webserver in certain environments.
  */
 public interface HelidonTaskExecutor extends Closeable {
