+++ nima/webserver/webserver/src/main/java/io/helidon/nima/webserver/ServerListener.java
@@ -43,7 +43,6 @@ import io.helidon.common.socket.TlsSocket;
 import io.helidon.nima.common.tls.Tls;
 import io.helidon.nima.webserver.http.DirectHandlers;
 import io.helidon.nima.webserver.spi.ServerConnectionSelector;
-import io.helidon.nima.webserver.task.spi.HelidonTaskExecutor;
 
 import static java.lang.System.Logger.Level.DEBUG;
 import static java.lang.System.Logger.Level.ERROR;
