--- nima/webserver/webserver/src/main/java/io/helidon/nima/webserver/task/spi/package-info.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (c) 2023 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Helidon executor SPI.
- *
- * @see io.helidon.nima.webserver.ThreadPerTaskExecutor
- */
-package io.helidon.nima.webserver.task.spi;
